Updated House Goat
Updated Description From "[Mystic](/does/Mystic) x [Magma](/bucks/Magma) Doeling\n\nWe haven't yet decided which of the two doelings that we are going to keep out of [Mystic](/does/Mystic), but we do plan to retain one\n\nHouse Goat (Further abbreviated to HG) was rejected by Mystic, and had to be brought into the house for the first few days of her life to keep her warm and so that we could feed her. Even though she's out in the barn now, we still call her HG and she's always the first to run up to you and say \"hi\" when you enter the barn" To "[Mystic](/does/Mystic) x [Magma](/bucks/Magma) Doeling\n\nHouse Goat (Further abbreviated to HG) was rejected by Mystic, and had to be brought into the house for the first few days of her life to keep her warm and so that we could feed her. Even though she's out in the barn now, we still call her HG and she's always the first to run up to you and say \"hi\" when you enter the barn"