Updated Stayawhile HC Forged Steel
Set Description To "Unfortunately, Steel was unable to handle the heat wave that we experienced and peacefully passed away June 10th, 2025. We are looking forward to seeing how the few kids that we got out of him do.\n\nHis dam is on the Elite Does list and has been on the ADGA Top Ten list 3 years. Steel is very dairy and so far, has put udders on his daughters. We are excited to mix his genetics into our farm."