Updated Lil Pilchuck Creek Ameth Jewel
Set Description To "Jewel was our first keeper doe to give us an idea of what conformation runs in the herd. Although she was maturing into a decent doe, she hated being milked and shown, so when we needed to downsize our two-year-olds she was the first to go. She now lives on a pet farm with a family that is going to give her a life she enjoys."